bluetooth: bluez: GetName uses 'name' instead of 'alias'.
BlueZ produces an alias that includes the MAC address of
a remote device when the device name is unknown. Client
code of device/bluetooth must be able to detect an empty
name and also protect private information such as nearby
MAC addresses from being shared.

This patch changes our implementation to use the BlueZ
provided 'name' instead of 'alias'.

BUG=586438